Extract new folder row and change blur behavior
